custom hook for height/resize logic to simplify component
